add balance to main form

+++ car-selling/Domain/Operation.cs
@@ -7,11 +7,11 @@ public class Operation
 
     public DateTime Timestamp { get; set; }
     public string Description { get; set; }
-    public decimal Amount { get; set; }
+    public int Amount { get; set; }
 
     public Operation() { }
 
-    public Operation(DateTime timestamp, string description, decimal amount)
+    public Operation(DateTime timestamp, string description, int amount)
     {
         this.Timestamp = timestamp;
         this.Description = description;
